+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TestTigerProxyTls.java
@@ -601,7 +601,6 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
 
       assertThat(response.code()).isEqualTo(666);
     }
-    ;
   }
 
   @Test
@@ -900,6 +899,51 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
                     "src/test/resources/rsa.p12;00"))
         .isInstanceOf(TlsFatalAlert.class)
         .hasRootCauseInstanceOf(CertPathBuilderException.class);
+  }
+
+  @Test
+  void reverseProxyWithMultipleServersAlternativeNames_shouldSelectCorrectCertificate() throws UnirestException {
+    spawnTigerProxyWith(
+            TigerProxyConfiguration.builder()
+                    .tls(
+                            TigerTlsConfiguration.builder()
+                                    .serverIdentities(
+                                            List.of(
+                                                    new TigerConfigurationPkiIdentity("src/test/resources/epa-as-1.dev.epa4all.de_NIST_X509.p12"),
+                                                    new TigerConfigurationPkiIdentity(
+                                                            "src/test/resources/epa-as-2.dev.epa4all.de_NIST_X509.p12"),
+                                                    new TigerConfigurationPkiIdentity(
+                                                            "src/test/resources/erp-dev.app.ti-dienste.de_NIST_X509.p12"),
+                                                    new TigerConfigurationPkiIdentity(
+                                                            "src/test/resources/erp-ref.app.ti-dienste.de_NIST_X509.p12"),
+                                                    new TigerConfigurationPkiIdentity(
+                                                            "src/test/resources/idp-ref.zentral.idp.splitdns.ti-dienste.de_NIST_X509.p12"),
+                                                    new TigerConfigurationPkiIdentity(
+                                                            "src/test/resources/localhostIdentity.p12")))
+
+                                    .allowGenericFallbackIdentity(true)
+                                    .build())
+                    .proxyRoutes(
+                            List.of(
+                                    TigerConfigurationRoute.builder()
+                                            .from("/")
+                                            .to("http://localhost:" + fakeBackendServerPort + "/foobar")
+                                            .hosts(List.of("epa-as-1.dev.epa4all.de"))
+                                            .build(),
+                                    TigerConfigurationRoute.builder()
+                                            .from("/")
+                                            .to("http://localhost:" + fakeBackendServerPort + "/deep/foobar")
+                                            .hosts(List.of("epa-as-2.dev.epa4all.de"))
+                                            .build(),
+                                    TigerConfigurationRoute.builder()
+                                            .from("/")
+                                            .to("http://localhost:" + fakeBackendServerPort + "/vauServer")
+                                            .hosts(List.of("erp-ref.zentral.erp.splitdns.ti-dienste.de", "erp-ref.app.ti-dienste.de", "subscription-ref.zentral.erp.splitdns.ti-dienste.de"))
+                                            .build()))
+                    .build());
+
+    executeRequestToPathWhileOnlyTrusting(
+            "erp-ref.zentral.erp.splitdns.ti-dienste.de", "src/test/resources/erp-ref.app.ti-dienste.de_NIST_X509.p12;00");
   }
 
   @SneakyThrows